Add files via upload
In this project myself and my group mates created a series of 20+ honeypots that resemble different aspects of a web server and a network. With an end goal of getting inside the mind of an attacker and learning how they think/operate and learning the numerous amount of vulnerabilities we were able to capture an enormous amount of data to analyze and learn from.